pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-wtf>=1.2.2",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "routes>=2.5.1",
    "sqlalchemy>=2.0.38",
    "scraper>=0.1.0",
    "werkzeug>=3.1.3",
    "instaloader>=4.14.1",
    "moviepy>=2.1.2",
    "pillow>=10.4.0",
    "uploader>=0.5.1",
    "instagrapi>=2.1.3",
    "numpy>=2.2.3",
    "opencv-python>=4.11.0.86",
]
